+++ packages/koot-webpack/loaders/koot-dev-client.js
@@ -13,17 +13,24 @@ module.exports = function(source) {
 
     if (!__react) return source;
 
+    const expExportDefault = `export default _default`;
     source = source.replace(
         /(var _default = .+?;\n*)(;\n\n\(function \(\) \{\n[ ]*var reactHotLoader = )/gm,
-        `$1\n/* harmony default export */ __webpack_exports__["default"] = (_default)$2`
+        // `$1\n/* harmony default export */ __webpack_exports__["default"] = (_default)$2`
+        `$1\n${expExportDefault}$2`
     );
 
-    if (!source.includes('__webpack_exports__["default"]')) {
-        const match = /var\s+_default\s*=\s*(.+?);/.exec(source);
+    // if (!source.includes('__webpack_exports__["default"]')) {
+    //     const match = /var\s+_default\s*=\s*(.+?);/.exec(source);
+    //     if (Array.isArray(match) && match.length > 1) {
+    //         source += `\n/* harmony default export */ __webpack_exports__["default"] = _default;\n`;
+    //     }
+    // }
+
+    if (!source.includes(expExportDefault)) {
+        const match = /var\s+_default\s*=\s*.+?;/.exec(source);
         if (Array.isArray(match) && match.length > 1) {
-            source += `\n/* harmony default export */ __webpack_exports__["default"] = ${
-                match[1]
-            };\n`;
+            source += `\n${expExportDefault};\n`;
         }
     }