+++ test/projects/standard/src/components/_layout/main/index.jsx
@@ -1,9 +1,7 @@
-import React from 'react'
-import { extend } from 'koot'
+import React from 'react';
+import { extend } from 'koot';
 
 export default extend({
     styles: require('./styles.less'),
     name: 'Main'
-})(
-    props => <div {...props} />
-)
+})(props => <div id="main" {...props} />);