+++ CHANGELOG.md
@@ -47,12 +47,18 @@
         -   `url-loader` -> _2.0.1_
         -   `os-locale` -> _4.0.0_
     -   minor
+        -   `@babel/core` -> _7.5.0_
+        -   `@babel/plugin-proposal-class-properties` -> _7.5.0_
+        -   `@babel/plugin-proposal-object-rest-spread` -> _7.5.0_
+        -   `@babel/plugin-transform-runtime` -> _7.5.0_
+        -   `@babel/preset-env` -> _7.5.0_
+        -   `acorn` -> _6.2.0_
         -   `cli-spinners` -> _2.2.0_
         -   `fs-extra` -> _8.1.0_
         -   `inquirer` -> _6.4.1_
         -   `mini-css-extract-plugin` -> _0.7.0_
         -   `open` -> _6.4.0_
-        -   `react-hot-loader` -> _4.12.2_
+        -   `react-hot-loader` -> _4.12.3_
         -   `webpack` -> _4.35.2_
         -   `webpack-dev-server` -> _3.7.2_
     -   patch