+++ CHANGELOG.md
@@ -58,7 +58,7 @@
         -   `inquirer` -> _6.4.1_
         -   `mini-css-extract-plugin` -> _0.7.0_
         -   `open` -> _6.4.0_
-        -   `react-hot-loader` -> _4.12.3_
+        -   `react-hot-loader` -> _4.11.2_
         -   `webpack` -> _4.35.2_
         -   `webpack-dev-server` -> _3.7.2_
     -   patch