+++ CHANGELOG.md
@@ -30,6 +30,7 @@
     -   分析模式
         -   优化分析模式输出文件名的可读性
 -   添加依赖包
+    -   `@babel/preset-typescript`
     -   `@types/node`
     -   `@types/react`
     -   `@types/react-dom`
@@ -52,13 +53,12 @@
         -   `@babel/plugin-proposal-object-rest-spread` -> _7.5.0_
         -   `@babel/plugin-transform-runtime` -> _7.5.0_
         -   `@babel/preset-env` -> _7.5.0_
-        -   `acorn` -> _6.2.0_
         -   `cli-spinners` -> _2.2.0_
         -   `fs-extra` -> _8.1.0_
         -   `inquirer` -> _6.4.1_
         -   `mini-css-extract-plugin` -> _0.7.0_
         -   `open` -> _6.4.0_
-        -   `react-hot-loader` -> _4.11.1_
+        -   `react-hot-loader` -> _4.12.3_
         -   `webpack` -> _4.35.2_
         -   `webpack-dev-server` -> _3.7.2_
     -   patch